Level design carter (#41)
* Create Carter Hopkins.txt

* Added Basic Level + Scene Swap Script

Added the basic level as well as a simple player cube to run around it with. Simple flat color textures are included that are to be replaced. Bounding walls around the level are to be made invisible later by turning off the mesh renderer.

Additionally included is a Scene Transition Script that swaps from the LevelDesignTestScene1 and LevelDesignTestScene2 by hitting 1 and 2.

I still have more work Id like to do to clean things up and make it pretty, but I wanted to push this so theres at least a basic level present that people could use to scale and test things in.

* Fixing Minor issue

* Fixing Terrain Name

* Fixed more minor issues