Added Smoothness and Shake (#58)
Camera moves more smoothly rather than abruptly. Also added a coroutine
for camera shake people can use. In the example, I made it so that if
you hit one of the large objects the camera shakes just to demonstrate.
Coroutine also takes duration and magnitude parameters so that people
can decide how much they want and for how long when they call it.