Input Utilities - Christos Othonos (#71)
Gather round dear reviewers, for this is the marvellous **InputUtils**!!
Read on and be amazed by this magical utility!

What this is:
InputUtils is a public static script that can be used by any script in
our project. It isn't a MonoBehavior so it doesn't need to be added to
the scene to work its magic. Furthermore its not a ScriptableObject, so
it doesn't need to be added as an asset. It works like a 'utility'
singleton similar to the Input system.

What it does:
It offers a consistent interface to check for game inputs. It also
contains useful utilities like groundPoint which returns the world
position of the mouse when projected to the ground plane.

Why its useful:
It stores string constants for the Input action names in one place. This
way, we reduce the amount of errors and fixes should the names of the
inputs suddenly change. Furthermore, it allows the reuse of input
related boiler plate code, thus reducing bloat in our game scripts.

Downsides:
The readonly strings have to be changed inside the code. This causes a
re-compilation which can (understandably) annoy some people. Arguably a
small price to pay for the utility provided by this _amazing_ script!