Created a Mini Map (#92)
# Summary
Mini Map Appears in the game that gives a higher view of the level and
follows the player. Put a green orb to represent the player but will
also have to add it to enemies and obstacles eventually. Also made the
second camera with effects that can make the orbs glow. Map camera can
only see orbs and anything on the map layer and the main camera can not
see the orbs.

# Details

# Demo Explanation
Go to the main level scene and push play.

# Bug Fixes

# Deleted

---------

Co-authored-by: cothon01 <christos.othonos.stratos@gmail.com>